requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pyyaml>=6.0.1
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
openpyxl>=3.1.0
pymupdf>=1.23.0
